requirements.txt
# GeoSpark Requirements

## Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4

## AI/ML Dependencies
openai==1.3.7
anthropic==0.7.8
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0

## Geospatial Dependencies
geopandas==0.14.1
shapely==2.0.2
folium==1.0.0
rasterio==1.3.9
pyproj==3.6.1

## Security Dependencies
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

## Information Retrieval
elasticsearch==8.11.0
chromadb==0.4.18
sentence-transformers==2.2.2

## Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

## Environment Management
python-dotenv==1.0.0
pydantic-settings==2.1.0

## HTTP Client
httpx==0.25.2
aiohttp==3.9.1

## Data Processing
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

## Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0